Migrate all callers to BazelWorkspace, delete free run_query(), merge query.py into workspace.py
- Migrate find_orphans.py, bazel_query.py, ci_decide.py, check_pytest_main.py
  to use BazelWorkspace.query() instead of free run_query()
- Inline run_query() logic into BazelWorkspace.query() method
- Delete util/bazel/query.py, merge BazelLabel and BazelWorkspace into
  util/bazel/workspace.py alongside existing env var helpers
- Update all imports (9 files) and BUILD deps from //util/bazel:query
  to //util/bazel:workspace
- Rename test_query.py to test_workspace.py
