+++ devinfra/BUILD.bazel
@@ -29,7 +29,7 @@ py_library(
     srcs = ["check_pytest_main.py"],
     imports = [".."],
     visibility = ["//visibility:public"],
-    deps = ["//util/bazel:query"],
+    deps = ["//util/bazel:workspace"],
 )
 
 # === Gazelle (Python BUILD file generator) ===
