+++ devinfra/orphans/BUILD.bazel
@@ -6,7 +6,6 @@ py_binary(
     imports = ["../.."],
     visibility = ["//:__subpackages__"],
     deps = [
-        "//util/bazel:query",
         "//util/bazel:workspace",
         "@pypi//pathspec",
         "@pypi//pygit2",
